example/dynamic/files/requirements.txt
fastapi
pycryptodome
gunicorn
uvicorn
